+++ airflow/contrib/operators/adls_list_operator.py
@@ -19,8 +19,8 @@
 
 from typing import Iterable
 
-from airflow.contrib.hooks.azure_data_lake_hook import AzureDataLakeHook
 from airflow.models import BaseOperator
+from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
 from airflow.utils.decorators import apply_defaults
