+++ airflow/models/connection.py
@@ -287,7 +287,7 @@ class Connection(Base, LoggingMixin):
             from airflow.hooks.docker_hook import DockerHook
             return DockerHook(docker_conn_id=self.conn_id)
         elif self.conn_type == 'azure_data_lake':
-            from airflow.contrib.hooks.azure_data_lake_hook import AzureDataLakeHook
+            from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
             return AzureDataLakeHook(azure_data_lake_conn_id=self.conn_id)
         elif self.conn_type == 'azure_cosmos':
             from airflow.providers.microsoft.azure.hooks.azure_cosmos import AzureCosmosDBHook
